refactor(proxy): separate request selection, credentials, and accounting
